Implementierung Lasso (sklearn)
