--- drevalpy/models/Lasso/__init__.py
@@ -1,1 +1,0 @@
-"""Module for the Lasso model."""
