--- drevalpy/models/Lasso/hyperparameters.yaml
@@ -1,7 +1,0 @@
-Lasso:
-  alpha:
-    - 0.0001
-    - 0.001
-    - 0.01
-    - 0.1
-    - 1
