+++ drevalpy/models/baselines/hyperparameters.yaml
@@ -208,4 +208,4 @@ Lasso:
     - 0.001
     - 0.01
     - 0.1
-    - 1 
+    - 1
