+++ tests/models/test_baselines.py
@@ -40,6 +40,7 @@ from drevalpy.models.drp_model import DRPModel
         "ProteomicsRandomForest",
         "ProteomicsElasticNet",
         "AdaBoostDecisionTree",
+        "Lasso",
     ],
 )
 @pytest.mark.parametrize("test_mode", ["LTO", "LPO", "LCO", "LDO"])
